Uglily migrate to Swoosh DataChecker
